Generate systemd instead of upstart configs
The service needs to be stopped when installing the new config files,
otherwise the service will still be active, but report "Loaded: not-found".
See https://github.com/twingly/capistrano-twingly/issues/31#issuecomment-486559472

For this to work we also need to make sure the deploy user is allowed to run
these commands on the appservers.

Related to #45