Automatically lookup which server manager is used on each server
This way we don't have to use different SRV records for ubuntu 14
and ubuntu 18 (idea from comments in #45).

We will always have a "failure" in the output on ubuntu 14 for the
command that checks if systemd is installed, but I found no way to
get around that. Part of the output displayed when deploying:

    00:03 deploy:lookup_server_service_manager
          01 command -v systemctl
          01 /bin/systemctl
        ✘ 01 deploy@localhost 0.003s
        ✔ 01 deploy@localhost 0.004s