Make sure systemd target is loaded before trying to stop it
During the initial deploy there is no systemd target to load which
means the command `systemctl stop` results in an error.

Now we check that the target exists before running `systemctl stop`.

The reason for stopping it in the first place is to make sure no
old processes are left behind, which could otherwise happend if
we remove/rename one of the processes in the procfile.
See https://github.com/twingly/capistrano-twingly/issues/31#issuecomment-486578861

close #49