Add workflow for GitHub actions
I used the action setup-ruby to configure and install Ruby for the workflow.
It was pretty straightforward to configure matrices to accomplish the
same ideas we have going for Travis, there are a lot of examples to follow.
https://github.com/ruby/setup-ruby

It can be a bit confusing since there is a deprecated variant if you
don't pay attention (like I didn't)
https://github.com/marketplace/actions/setup-ruby-environment-with-ruby-build

Unfortunately there is no way of doing a "allow-failure" like in Travis CI.
There is an ongoing thread however where people request it.
Seems like we are not the only ones migrating.
https://github.com/actions/toolkit/issues/399