Use net-ssh ~> 7.0
The deployment of norrstrom failed:

```
Caused by:
OpenSSL::PKey::PKeyError: rsa#set_key= is incompatible with OpenSSL 3.0
```

This was fixed in net-ssh 7, according to: https://stackoverflow.com/a/72075830.

The latest versions of net-ssh are not yet included in the changelog, but I used this new version of capistrano-twingly in the test repo and the deployment worked fine. Since we use capistrano in the same way in all our projects, this test should suffice.