Bump version to 4.1.0
Related to the change in #69.

That PR adds a fallback DNS server, and I thought of that as a new
feature. One could also think of this as a bugfix, but the fact that
your local DNS server is unable to find the record isn't really a bug
in this gem.