Make it possible to set X-Forwarded-Proto header to "https"
This header is apparently needed if we forward a HTTPS request, see #41.

I didn't bother to make this work for all possible headers we might
want to set in the future. We don't use the nginx functionality of this
gem that much, and hopefully we'll be able to switch to a better type of
deployment sometime in the future anyway.

close #41